BDOG-3888: tidy up test. Move out some deps to support class
